// frontend/src/types/aas.ts
export enum ReferenceTypes {
  ExternalReference = 'ExternalReference',
  ModelReference = 'ModelReference',
}
// ...
export interface Key {
  type: string;
  value: string;
}

export interface Reference {
  type: ReferenceTypes;
  keys: Key[];
  referredSemanticId?: Reference;
}